requirements.txt
RPi.GPIO
Flask
atexit